added app.js file
